Made changes to accomodate memory mgmt gcalloc
