+++ utils.ts
@@ -30,18 +30,19 @@ export function PyValue(typ: Type, result: number, mem: any): Value {
       var full_string = "";
       while (i < string_length) {
         let ascii_val = view[data / 4 + i];
+        console.log("ASCII value "+ascii_val);
         var char = String.fromCharCode(ascii_val);
         full_string += char;
         i += 1;
       }
       //new
       var i=0;
-      while(i<30){
+      while(i<5){
         console.log("View "+i+" Value "+view[i]);
         i+=1;
       }
       //old
-      console.log("Full string "+full_string.length);
+      console.log("Full string "+full_string);
       return PyString(full_string, result);
     case "number":
       // console.log("Actual length "+result);
